+++ Quake.java
@@ -1,6 +1,6 @@
 public class Quake extends ActionedEntity{
     private int animationRate;
-    public Quake(String name, Point position, double rate, int animationRate){
+    public Quake(String name, Point position, int rate, int animationRate){
         super(name, position, rate, "quake");
         this.animationRate = animationRate;
     }
